% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getmm.R
\name{getmm}
\alias{getmm}
\title{Get datasets from Mastering 'Metrics}
\usage{
getmm(data, envir = parent.frame())
}
\arguments{
\item{data}{A string specifying the name of a Mastering 'Metrics dataset.}

\item{envir}{An environment in which to store the object. Default is the parent environment.}
}
\value{
A dataframe.
}
\description{
Get datasets from Mastering 'Metrics
}
\examples{
getmm("mlda")
}

