% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reg.R
\name{reg}
\alias{reg}
\title{Run a OLS regression just like running \code{reg y x, r} in Stata.}
\usage{
reg(formula, data, ...)
}
\arguments{
\item{formula}{Regression formula.}

\item{data}{A dataframe.}

\item{...}{Extra arguments.}
}
\value{
A dataframe that includes regression coefficients and robust standard errors.
}
\description{
Run a OLS regression just like running \code{reg y x, r} in Stata.
}
\examples{
reg(mpg ~ hp + wt, data = mtcars)
}

